requirements.txt
cryptography==42.0.5
fastapi==0.110.0
httpx==0.27.0
langchain==0.1.12
langchain_community==0.0.28
langchain_core==0.1.31
langchain_openai==0.0.8
openai==1.14.0
pydantic==2.6.4
python-dotenv==1.0.1
SQLAlchemy==2.0.25
chromadb==0.4.22
beautifulsoup4==4.12.3
PyMySQL==1.1.0


 
